Move duration package to runtime/pkg
